refactor: BookingResponse DTO들을 BookingDetail로 분리
